requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
supabase==1.2.0
sqlmodel==0.0.8
httpx==0.25.0
pytest==7.4.3
flake8==6.1.0